This PR mainly focuses on optimizing `gfx-chain`. Other changes made:
* Rename `sync::Sync` to `sync::SyncData` to avoid overlapping with a core Rust type's name.
* Add `IntoIterator` implementations to all types in `schedule`.
* `gfx-chain` now handles the case where two adjacent links in a chain both contain multiple queues.
* The `wait_factor` mechanism is now fixed, and isn't universally `0` in all submissions.

Benchmarks before optimization
------------------------------
```
tiny   + single-queue:  67899 iters ran in average 0.047 ± 0.010 ms (0.28% ± 0.06% of 16.6 ms)
tiny   + multi-queue :  43391 iters ran in average 0.080 ± 0.018 ms (0.48% ± 0.11% of 16.6 ms)
small  + single-queue:   6456 iters ran in average 0.560 ± 0.091 ms (3.37% ± 0.55% of 16.6 ms)
small  + multi-queue :   3458 iters ran in average 1.094 ± 0.232 ms (6.59% ± 1.40% of 16.6 ms)
medium + single-queue:   1189 iters ran in average 3.181 ± 0.453 ms (19.16% ± 2.73% of 16.6 ms)
medium + multi-queue :    615 iters ran in average 6.313 ± 0.877 ms (38.03% ± 5.28% of 16.6 ms)
large  + single-queue:     28 iters ran in average 142.581 ± 7.033 ms (858.92% ± 42.37% of 16.6 ms)
large  + multi-queue :     18 iters ran in average 221.229 ± 17.011 ms (1332.71% ± 102.48% of 16.6 ms)
huge   + single-queue:     12 iters ran in average 356.965 ± 11.768 ms (2150.39% ± 70.89% of 16.6 ms)
huge   + multi-queue :      8 iters ran in average 527.094 ± 28.118 ms (3175.27% ± 169.38% of 16.6 ms)
```

Benchmarks after optimization
-----------------------------
```
tiny   + single-queue: 124284 iters ran in average 0.021 ± 0.003 ms (0.12% ± 0.02% of 16.6 ms)
tiny   + multi-queue :  83506 iters ran in average 0.036 ± 0.005 ms (0.22% ± 0.03% of 16.6 ms)
small  + single-queue:  17860 iters ran in average 0.166 ± 0.016 ms (1.00% ± 0.10% of 16.6 ms)
small  + multi-queue :  11654 iters ran in average 0.285 ± 0.031 ms (1.71% ± 0.18% of 16.6 ms)
medium + single-queue:   5242 iters ran in average 0.594 ± 0.054 ms (3.58% ± 0.32% of 16.6 ms)
medium + multi-queue :   3489 iters ran in average 0.976 ± 0.080 ms (5.88% ± 0.48% of 16.6 ms)
large  + single-queue:    346 iters ran in average 10.206 ± 0.559 ms (61.48% ± 3.37% of 16.6 ms)
large  + multi-queue :    213 iters ran in average 17.310 ± 0.632 ms (104.27% ± 3.81% of 16.6 ms)
huge   + single-queue:    188 iters ran in average 19.098 ± 0.699 ms (115.05% ± 4.21% of 16.6 ms)
huge   + multi-queue :    122 iters ran in average 30.534 ± 1.285 ms (183.94% ± 7.74% of 16.6 ms)
```